docs(ui): fold --bui-bg-inherit story into colors stories
